move content components inside content folder
