+++ src/components/Content/ResultPanel/ResultPanel.css
