+++ src/components/Content/ResultPanel/ResultPanel.jsx
@@ -1,6 +1,6 @@
 import { useContext } from "react";
 
-import { GlobalContext } from "../../context/Global";
+import { GlobalContext } from "../../../context/Global";
 import './ResultPanel.css';
 
 const ResultPanel = () => {
